In java code I used Long variable for id, in  the database I use int for ids, so I decided to change Long for Integer in java code

+++ src/main/java/io/github/mat3e/LangRepository.java
@@ -10,12 +10,12 @@ class LangRepository {
 
      LangRepository(){
          languages = new ArrayList<>();
-         languages.add(new Lang(1L, "Hello", "en"));
-         languages.add(new Lang(2L, "Witam","pl"));
-         languages.add(new Lang(3L, "Hola", "es"));
+         languages.add(new Lang(1, "Hello", "en"));
+         languages.add(new Lang(2, "Witam","pl"));
+         languages.add(new Lang(3, "Hola", "es"));
      }
 
-     Optional<Lang> findById(Long id) {
+     Optional<Lang> findById(Integer id) {
          return languages.stream().filter(l -> l.getId().equals(id)).findFirst();
      }
 }
